environment_detector.py
import io
import json
import uuid
from typing import Dict, List, Any, Optional

from google.cloud import vision
from google.cloud import storage

# This would be the actual import when using the Gemini API
# from google.cloud import aiplatform
# from vertexai.generative_models import GenerativeModel


class EnvironmentDetector:
    """Main class for detecting environment elements in media"""

    def __init__(self, project_id: str, bucket_name: str):
        """Initialize the EnvironmentDetector with GCP credentials

        Args:
            project_id: Google Cloud Project ID
            bucket_name: Cloud Storage bucket for storing metadata
        """
        self.project_id = project_id
        self.bucket_name = bucket_name
        self.vision_client = vision.ImageAnnotatorClient()
        self.storage_client = storage.Client(project=project_id)
        # Initialize Gemini model - commented out as placeholder
        # aiplatform.init(project=project_id)
        # self.gemini_model = GenerativeModel('gemini-pro-vision')

    def detect_environment(self, image_path: str) -> Dict[str, Any]:
        """Detect environment elements in an image

        Args:
            image_path: Path to the image file

        Returns:
            Dict containing environment tags
        """
        # Load image
        with io.open(image_path, 'rb') as image_file:
            content = image_file.read()

        image = vision.Image(content=content)

        # Detect labels
        response = self.vision_client.label_detection(image=image)
        labels = response.label_annotations

        # Detect properties
        prop_response = self.vision_client.image_properties(image=image)
        properties = prop_response.image_properties_annotation

        # Process labels to identify environment elements
        environment_elements = []
        is_indoor = False
        is_outdoor = False

        for label in labels:
            if label.description.lower() in ['indoor', 'room', 'interior']:
                is_indoor = True
            elif label.description.lower() in ['outdoor', 'outside', 'exterior']:
                is_outdoor = True

            environment_elements.append({
                'name': label.description,
                'confidence': label.score,
                'category': self._classify_element(label.description)
            })

        # Use Gemini API for contextual understanding
        gemini_response = self._analyze_with_gemini(image_path, labels)

        # Construct environment metadata
        environment = {
            'setting': {
                'type': 'indoor' if is_indoor else 'outdoor' if is_outdoor else 'unknown',
                'confidence': 0.85  # Example confidence
            },
            'location': {
                'description': gemini_response.get('location', 'unknown'),
                'confidence': gemini_response.get('location_confidence', 0.7)
            },
            'elements': environment_elements,
            'conditions': {
                'lighting': gemini_response.get('lighting', 'unknown'),
                'weather': gemini_response.get('weather', 'unknown'),
                'timeOfDay': gemini_response.get('time_of_day', 'unknown'),
                'confidence': gemini_response.get('conditions_confidence', 0.7)
            }
        }

        return environment

    def _classify_element(self, label: str) -> str:
        """Classify environment element into categories"""
        furniture = ['chair', 'table', 'desk', 'sofa', 'bed', 'shelf']
        nature = ['tree', 'plant', 'flower', 'grass', 'mountain', 'sky', 'cloud']
        architecture = ['building', 'wall', 'door', 'window', 'ceiling', 'floor']

        label_lower = label.lower()

        if any(item in label_lower for item in furniture):
            return 'furniture'
        elif any(item in label_lower for item in nature):
            return 'nature'
        elif any(item in label_lower for item in architecture):
            return 'architecture'
        else:
            return 'other'

    def _analyze_with_gemini(self, image_path: str, vision_labels: List) -> Dict[str, Any]:
        """Use Gemini API for enhanced understanding of environment"""
        # This would be the actual implementation using Gemini API
        # with open(image_path, 'rb') as f:
        #     image_content = f.read()
        # response = self.gemini_model.generate_content([image_content, "Describe the environment in this image"])
        
        # Example simulated response structure
        return {
            'location': 'modern office space',
            'location_confidence': 0.82,
            'lighting': 'bright indoor lighting',
            'weather': 'not applicable',
            'time_of_day': 'daytime',
            'conditions_confidence': 0.75
        }

    def store_environment_metadata(self, environment: Dict[str, Any]) -> str:
        """Store environment metadata in Cloud Storage
        
        Args:
            environment: Environment metadata dictionary
            
        Returns:
            Unique media ID for the stored metadata
        """
        bucket = self.storage_client.bucket(self.bucket_name)
        media_id = str(uuid.uuid4())
        blob = bucket.blob(f'tags/{media_id}.json')
        
        blob.upload_from_string(
            json.dumps(environment),
            content_type='application/json'
        )
        
        return media_id

    def get_environment_metadata(self, media_id: str) -> Optional[Dict[str, Any]]:
        """Retrieve environment metadata from Cloud Storage
        
        Args:
            media_id: Unique ID for the media
            
        Returns:
            Environment metadata dictionary or None if not found
        """
        bucket = self.storage_client.bucket(self.bucket_name)
        blob = bucket.blob(f'tags/{media_id}.json')
        
        if not blob.exists():
            return None
            
        content = blob.download_as_string()
        return json.loads(content)

    def update_environment_metadata(self, media_id: str, tags: List[str], 
                              override: bool = False) -> bool:
        """Update environment metadata in Cloud Storage
        
        Args:
            media_id: Unique ID for the media
            tags: New or additional tags
            override: Whether to override existing tags
            
        Returns:
            True if update was successful, False otherwise
        """
        metadata = self.get_environment_metadata(media_id)
        
        if not metadata:
            return False
            
        if override:
            # Replace all elements with new tags
            metadata['elements'] = [{'name': tag, 'confidence': 1.0, 
                                   'category': self._classify_element(tag)} 
                                  for tag in tags]
        else:
            # Add new tags to existing elements
            existing_tags = {elem['name'].lower() for elem in metadata['elements']}
            for tag in tags:
                if tag.lower() not in existing_tags:
                    metadata['elements'].append({
                        'name': tag,
                        'confidence': 1.0,
                        'category': self._classify_element(tag)
                    })
        
        # Store updated metadata
        bucket = self.storage_client.bucket(self.bucket_name)
        blob = bucket.blob(f'tags/{media_id}.json')
        
        blob.upload_from_string(
            json.dumps(metadata),
            content_type='application/json'
        )
        
        return True