requirements.txt
flask==2.2.3
gunicorn==20.1.0
google-cloud-vision==3.4.0
google-cloud-storage==2.7.0
# For Gemini API when available
# google-cloud-aiplatform==1.25.0
requests==2.28.2
python-dotenv==1.0.0